requirements.txt
Django==2.1
#django-extensions==3.1.0
djangorestframework==3.10.3
pytz==2021.1
